Added image test in demo.
Image is drawn and saved as png. It need slight adjustment in the area where it is drawn. Remember TODO checklist.

Everything has been converted into `int` instead of `double`. It's much easier to use `int` since they'll be used like pixels, anyway.

`GenerateCellList` added in `CellGenerator`. It needs to be generated using "Park-Miller Normal Distribution" (which I can't find, besides Normal Gaussian Distribution. That's the one that was linked in the instruction as well). Currently, it is generated randomly.

Documentation is required after `CellGenerator` class is completed. It should  also explain how  things works there.