Step 4 is in progress.
I found a github repository that contains the implementation. It has the same license, thankfully. It is sort of 'rewritten' to fit the use case.

The implementation for the Delaunay Triangulation is in its own namespace. It should be noted that, when it is all done, only `TunnelGenerator` class should have access to it.