Step 4 is still in progress.
I'm making this commit because I'm about to delete a lot of stuff. I'm having difficult time doing this. Let's just make this a lot simpler by doing this manually.

Next time, I need to read what the algorithm is all about before giving up and look for existing implementation of that thing.

From what I see, in its simplest explanation, Delaunay Triangulation is simply creating a triangle that's connected between points that connects to the circle.

It should be sort of ... doable? I just need to calculate the circle and have it find its way to connect with each other. The only problem here happens if the circle is somehow connected to four lines.

I can only hope that rarely happens. Because if it does, that's going to be a problem.