Step 4 is done.
'couple of things. First, I found another implementation that's easier to use. I moved over to use delaunator-sharp made by nol1fe (https://github.com/nol1fe/delaunator-sharp). It has the same license (MIT) as this repository. It is placed in a separate namespace along with the license notice.

Also, some modifications are made there.

Second, there's a note. We need to find a better way to attach the new tunnels with the cells properly. Right now, it knows this based on the center location position of the cell and the point in edge. I've put it on TODO to fix it. (Note: Please actually fix it. I don't want to deal with this problem when it actually becomes a problem in the future.)

Note: The next phase is to use Minimum Spanning Tree on them. All adjustments on tunnels are in TunnelGenerator and the method is there to be implemented.