Refactoring and bug fixes.
The draw cell and line has been changed in the demo project. Now the cube also draws the coordinate of its top left and bottom right points at the top left corner. This is done to help debugging some stuff.

While refactoring, I found out that the results from flocking separation is apparently pivoted slightly by its own center. This was caught late because we increment the points with center to adjust the position of the stuff in canvas. Imagine my surprise when it broke the separation when I was testing the refactored code.

Anyway, it does explain why it the overlapping connection is a bit whack. It detects these 'ghost' cells and then deleted the line because of it.

Since it's pivoted, a temporary fix has been placed. Instead of incrementing the center at canvas, it was moved to right before the method returns the arranged cells. And thus, it fixed the problem in phase 4.