Added multiple simple methods.
The methods are for the grid section. They are made as an alternative for a faster processing time for points to grid conversion.

The methods are "CreateTunnelLayerSimple", which is a simpler version of "CreateTunnelLayer". They are more effective used on a straight line tunnel.

"VerifyConnection" is used to verify the tunnel connection, to ensure that it is connected to a tunnel.

"IntersectGrid" is used to change intersecting grid block to a given block type.

They are in the sync version of GridConverter. I've been thinking about deprecating the async version of GridConverter, GridConverterAsync, as it might not be needed anymore.

Also, removed couple of junk dependencies.